% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmclustering.R
\name{cluster.bic}
\alias{cluster.bic}
\title{HMM BIC}
\usage{
cluster.bic(hmm, eps = 0.001)
}
\arguments{
\item{hmm}{An \code{hmm} object (see \code{\link[seqHMM]{build_hmm}}).}

\item{eps}{A threshold epsilon for counting parameters.}
}
\value{
The BIC of \code{hmm}.
}
\description{
Compute the BIC of a single HMM given a threshold epsilon for counting
parameters. Auxiliary function used in \code{\link{size.search}}.
}
\seealso{
Used in \code{\link{size.search}}.
}
