% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmclustering.R
\name{count.parameters}
\alias{count.parameters}
\title{Count HMM Parameters}
\usage{
count.parameters(hmm, eps = 0.001)
}
\arguments{
\item{hmm}{An \code{hmm} object (see \code{\link[seqHMM]{build_hmm}}).}

\item{eps}{A threshold epsilon for counting parameters.}
}
\value{
The number of parameters larger than \code{eps}.
}
\description{
Count the number of parameters in an HMM larger than a small number epsilon.
Auxiliary function used in \code{\link{partition.bic}} and
\code{\link{cluster.bic}}.
}
\seealso{
Used in \code{\link{partition.bic}} and \code{\link{cluster.bic}}.
}
