% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{emission.heatmap}
\alias{emission.heatmap}
\title{Heatmap Emission Probabilities}
\usage{
emission.heatmap(emission, base_size = 10)
}
\arguments{
\item{emission}{A matrix with emission probabilities (see also
\code{\link[seqHMM]{build_hmm}}).}

\item{base_size}{Numerical, a size parameter for the plots made using \code{ggplot2}
(see \link[ggplot2]{theme}), default = 10.}
}
\description{
Plots a heatmap of an HMM's emission probabilities.
}
\seealso{
See \code{\link{hmm.clust}} for an example.
}
