% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmclustering.R
\name{partition.bic}
\alias{partition.bic}
\title{Partition BIC}
\usage{
partition.bic(partition, eps = 0.001)
}
\arguments{
\item{partition}{A list object with the partition of HMMs, a mixture of HMMs.}

\item{eps}{A threshold epsilon for counting parameters in
\code{\link{count.parameters}}.}
}
\value{
The BIC of the partition.
}
\description{
Compute the BIC of a partition given a threshold epsilon for counting
parameters. Auxiliary function used in \code{\link{hmm.clust}}.
}
\seealso{
Used in main function for the DBHC algorithm
  \code{\link{hmm.clust}}.
}
