% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmclustering.R
\name{smooth.probabilities}
\alias{smooth.probabilities}
\title{Smooth Probabilities}
\usage{
smooth.probabilities(probs, smoothing = 1e-04)
}
\arguments{
\item{probs}{A vector of raw probabilities.}

\item{smoothing}{Smoothing parameter for absolute discounting.}
}
\value{
A vector of smoothed probabilities.
}
\description{
Smooth a vector of probabilities using absolute discounting. Auxiliary
function used in \code{\link{smooth.hmm}}.
}
\seealso{
Used in \code{\link{smooth.hmm}}.
}
