% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{transition.heatmap}
\alias{transition.heatmap}
\title{Heatmap Transition Probabilities}
\usage{
transition.heatmap(transition, initial = NULL, base_size = 10)
}
\arguments{
\item{transition}{A matrix with transition probabilities (see also
\code{\link[seqHMM]{build_hmm}}).}

\item{initial}{An (optional) vector of initial probabilities.}

\item{base_size}{Numerical, a size parameter for the plots made using \code{ggplot2}
(see \link[ggplot2]{theme}), default = 10.}
}
\description{
Plots a heatmap of an HMM's initial and transition probabilities.
}
\seealso{
See \code{\link{hmm.clust}} for an example.
}
